+++ src/StorageSync/StorageSync/Common/StorageSyncConstants.cs
@@ -104,15 +104,6 @@ namespace Microsoft.Azure.Commands.StorageSync.Common
         public const string CloudTieringOff = "off";
 
         /// <summary>
-        /// The Offline Data Transfer on
-        /// </summary>
-        public const string OfflineDataTransferOn = "on";
-        /// <summary>
-        /// The Offline Data Transfer off
-        /// </summary>
-        public const string OfflineDataTransferOff = "off";
-
-        /// <summary>
         /// The afs agent registry key
         /// </summary>
         public const string AfsAgentRegistryKey = @"SOFTWARE\Microsoft\Azure\StorageSync\Agent";